requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.7.1
dj-database-url==2.1.0
Django==4.2.5
django-ckeditor==6.7.0
django-js-asset==2.1.0
django-on-heroku==1.1.2
django-phonenumber-field==7.2.0
gunicorn==21.2.0
idna==3.6
oauthlib==3.2.2
packaging==23.2
phonenumbers==8.13.21
Pillow==10.0.1
psycopg2==2.9.7
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.1.0
whitenoise==6.6.0
